% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{print.ccnlcob_result}
\alias{print.ccnlcob_result}
\title{Stampa sintetica di un risultato ccnlcob_result}
\usage{
\method{print}{ccnlcob_result}(x, ...)
}
\arguments{
\item{x}{Oggetto di classe \code{ccnlcob_result}.}

\item{...}{Ignorato; presente per coerenza con il generico \code{\link[=print]{print()}}.}
}
\value{
\code{x}, restituito invisibilmente.
}
\description{
Mostra i metadati principali e le dimensioni delle tabelle contenute in un
oggetto prodotto da \code{\link[=analyze_ccnl]{analyze_ccnl()}}.
}
\examples{
\dontrun{
res <- analyze_ccnl(cob_esempio)
print(res)
}
}
\seealso{
Other orchestrazione: 
\code{\link{analyze_ccnl}()},
\code{\link{write_results}()}
}
\concept{orchestrazione}
