% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tipologie_contrattuali}
\alias{tipologie_contrattuali}
\title{Tipologie contrattuali MLPS con macro-classe}
\format{
Un \code{data.table} con una riga per codice di tipologia contrattuale
e le colonne:
\describe{
\item{cod_tipologia_contrattuale}{character. Codice MLPS della tipologia
(es. \code{A.01.00}).}
\item{des_tipologia_contrattuale}{character. Descrizione ufficiale MLPS
della tipologia.}
\item{macro_tipologia}{character. Macro-classe di analisi, una fra:
\verb{Tempo indeterminato}, \verb{Tempo determinato}, \code{Apprendistato},
\code{Somministrazione}, \code{Intermittente}, \code{Collaborazioni}, \code{Tirocinio},
\code{Domestico}, \code{Altro}.}
\item{esclusa_standard}{logical. \code{TRUE} per le tipologie escluse dal
perimetro "standard" di \code{cnelR} (\code{C.01.00}, \code{B.04.00}, \code{B.03.00},
\code{A.04.00}, \code{A.04.01}).}
}
}
\source{
Ministero del Lavoro e delle Politiche Sociali, "Classificazioni
Standard" delle Comunicazioni Obbligatorie, foglio ST-TIPO CONTRATTI,
Rev.093 del 2026-04-26.
}
\usage{
tipologie_contrattuali
}
\description{
Lookup delle tipologie contrattuali delle Comunicazioni Obbligatorie
(classificazioni standard del Ministero del Lavoro e delle Politiche
Sociali) con la macro-classe usata da \code{\link[=classify_tipologia]{classify_tipologia()}} e
\code{\link[=ccnl_by_tipologia]{ccnl_by_tipologia()}} e il flag di esclusione dal perimetro "standard"
adottato da \code{cnelR}.
}
\examples{
head(tipologie_contrattuali)
}
\seealso{
\code{\link[=classify_tipologia]{classify_tipologia()}}, \code{\link[=ccnl_by_tipologia]{ccnl_by_tipologia()}}
}
\keyword{datasets}
